refactor(db-structure): rename Drizzle parser files to camelCase
  - Rename ast-utils.ts to astUtils.ts
  - Rename column-parser.ts to columnParser.ts
  - Rename enum-parser.ts to enumParser.ts
  - Rename expression-parser.ts to expressionParser.ts
  - Rename main-parser.ts to mainParser.ts
  - Rename relation-parser.ts to relationParser.ts
  - Rename table-parser.ts to tableParser.ts
  - Update all import statements to use new camelCase file names
