+++ frontend/packages/db-structure/src/parser/drizzle/index.ts
@@ -1,1 +1,1 @@
-export { processor } from './main-parser.js'
+export { processor } from './mainParser.js'
